% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster_coordinates}
\alias{cluster_coordinates}
\title{Cluster coordinates}
\format{
A data.frame with columns `cluster`, `latitude`, `longitude`.
}
\description{
Latitude and longitude of the maritime clusters used in the network.
}
