% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mardist.R
\name{distance_route}
\alias{distance_route}
\title{Distance calculation}
\usage{
distance_route(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{numeric. Longitude of the origin (decimal degrees).}

\item{lat1}{numeric. Latitude  of the origin (decimal degrees).}

\item{lon2}{numeric. Longitude of the destination (decimal degrees).}

\item{lat2}{numeric. Latitude  of the destination (decimal degrees).}
}
\value{
The distance in nautical miles (numeric).
}
\description{
Calculates the maritime distance between two geographic points.
}
\examples{
distance_route(121.48, 31.2198, 79.85, 6.95)
}
