% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mardist.R
\name{dtHaversine}
\alias{dtHaversine}
\title{Haversine great circle distance}
\usage{
dtHaversine(lat_from, lon_from, lat_to, lon_to, r = 6378137)
}
\arguments{
\item{lat_from}{numeric. Latitude of the *from* point (decimal degrees).}

\item{lon_from}{numeric. Longitude of the *from* point (decimal degrees).}

\item{lat_to}{numeric. Latitude of the *to* point (decimal degrees).}

\item{lon_to}{numeric. Longitude of the *to* point (decimal degrees).}

\item{r}{numeric. Radius of the sphere used for the calculation.
The default value 6378137  is the WGS84 equatorial radius in
metres.}
}
\value{
The distance between the two points, expressed in the same unit
        as r (metres by default).
}
\description{
Computes the great circle distance between two points on the Earth
surface using the Haversine formula.  The result is returned in the
same linear unit as the radius r that you supply (the default is the
WGS 84 Earth radius in metres).
}
\examples{
## Distance between Shanghai (121.48 E, 31.2198 N) and
## Goa (79.85 E, 6.95 N) the value is given in metres.
dtHaversine(31.2198, 121.48, 6.95, 79.85)
}
