% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mardist.R
\name{multi_point_route}
\alias{multi_point_route}
\title{Map a continues route that passes through specific points}
\usage{
multi_point_route(clist)
}
\arguments{
\item{clist}{A data.frame or data.table that must contain **two**
columns with the exact names `longitude` and `latitude`.  Any additional
columns are ignored.}
}
\value{
An object of class **`leaflet`** that displays all of the
  shortest maritime routes contained in `clist`.
}
\description{
Given a data frame / data.table (`clist`) that contains a list of
coordinates, the function draws a **leaflet** map with the
shortest route line that connects the points in thelist.  The function is useful for
visualising a full route that passes from specific points (e.g. all vessel port calls
during a certain period).
}
\examples{
## A multi-port route:
routes <- data.frame(cbind(
longitude=c(120,122,90,60,30,25,0,8),
latitude=c(30,31,40,32,40,34,34,55)))
multi_point_route(routes)

}
