% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{Maritime network}
\format{
A data.frame with N rows and 3 columns.
}
\source{
JRC
}
\description{
An igraph edge list that describes the global maritime network.
The object is a data frame with columns `cluster`, `lag_cluster`,
and `distance` (in metres).
}
