% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_wrapper.R
\name{route_map}
\alias{route_map}
\title{Map a single maritime route (convenient short name)}
\usage{
route_map(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{longitude of the origin (decimal degrees)}

\item{lat1}{latitude of the origin (decimal degrees)}

\item{lon2}{longitude of the destination (decimal degrees)}

\item{lat2}{latitude of the destination (decimal degrees)}
}
\value{
A `leaflet` map object.
}
\description{
Wrapper that forwards to `map_route`.  It exists only so that the
examples can use the short name `route_map`.
}
